+++ app/src/main/java/com/terra/planet/MainActivity.kt
@@ -750,6 +750,6 @@ class MainActivity : Activity() {
     }
 
     companion object {
-        const val VERSION = "0.15.2"
+        const val VERSION = "0.15.3"
     }
 }